Preview and Extra refactoring
