+++ app/src/main/java/space/taran/arknavigator/mvp/view/item/FileItemViewHolder.kt
@@ -4,9 +4,10 @@ import androidx.recyclerview.widget.RecyclerView
 import space.taran.arknavigator.R
 import space.taran.arknavigator.databinding.ItemFileGridBinding
 import space.taran.arknavigator.mvp.model.repo.index.ResourceMeta
-import space.taran.arknavigator.mvp.model.repo.index.ResourceMetaExtra
-import space.taran.arknavigator.ui.fragments.preview.PreviewAndThumbnail
-import space.taran.arknavigator.utils.*
+import space.taran.arknavigator.mvp.model.repo.preview.PreviewAndThumbnail
+import space.taran.arknavigator.ui.extra.ExtraLoader
+import space.taran.arknavigator.utils.ImageUtils
+import space.taran.arknavigator.utils.extension
 import java.nio.file.Path
 
 class FileItemViewHolder(private val binding: ItemFileGridBinding) :
@@ -28,12 +29,7 @@ class FileItemViewHolder(private val binding: ItemFileGridBinding) :
         val thumbnail = PreviewAndThumbnail.locate(path, resource)?.thumbnail
 
         ImageUtils.loadImageWithPlaceholder(thumbnail, placeholder, binding.iv)
-
-        ResourceMetaExtra.draw(
-            resource.kind,
-            resource.extra,
-            arrayOf(binding.primaryExtra, binding.secondaryExtra),
-            verbose = false)
+        ExtraLoader.load(resource, listOf(binding.primaryExtra, binding.secondaryExtra),  verbose = false)
     }
 
     override fun setText(title: String) = with(binding.root) {
